2-strlen.c

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,7 +6,7 @@
  *
  * Return: The length of @str.
  */
-size_t _strlen(const char *s)
+int _strlen(const char *s)
 {
 size_t length = 0;
 while (*s++)
